backend/foodgram_api/requirements.txt
django==4.0.6
djangorestframework==3.12.4
django-filter==22.1
pytz==2022.1
djoser==2.1.0

pillow==9.2.0
drf-extra-fields==3.4.0

psycopg2-binary==2.9.3
gunicorn==20.0.4
sqlparse==0.4.2

requests==2.26.0
PyJWT==2.1.0
